restructure codebase for better organization
