+++ frontend/src/app/events/EventPopupCard.tsx
