+++ frontend/src/app/events/page.tsx
@@ -1,16 +1,13 @@
 "use client";
 
 import React, { useEffect, useState } from "react";
-import { Calendar, MapPin, Clock, ExternalLink, User, Tag } from "lucide-react";
-import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card";
-import { CalendarComponent } from "@/components/calendar/Calendar";
-import { SlidingCard } from "@/components/ui/EventPopupCard";
+import { CalendarComponent } from "@/app/events/calendar/Calendar";
+import { SlidingCard } from "@/app/events/EventPopupCard";
 import { Boxes } from "@/components/ui/background-boxes";
-import { CardContainer, CardBody, CardItem } from "@/components/ui/3d-card";
-import Marquee from "@/components/ui/marquee";
 import { motion } from "framer-motion";
 import { config } from "@/config";
-import { Event } from "@/components/calendar/types";
+import { Event } from "@/app/events/calendar/types";
+import MarqueeSection from "./MarqueeSection";
 import "./Events.css";
 
 export default function Events() {
@@ -118,64 +115,7 @@ export default function Events() {
             </motion.div>
 
             {/* Marquee Section */}
-            <motion.div
-              className="marquee-section"
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              transition={{ delay: 0.4 }}
-            >
-              <Marquee pauseOnHover>
-                <div className="marquee-card-container">
-                  {events.map((event, index) => (
-                    <div key={index} className="marquee-card">
-                      {/* Background Image Section */}
-                      <div
-                        className="image-container"
-                        style={{
-                          backgroundImage: `url(${
-                            event.cover_image_url || ""
-                          })`,
-                        }}
-                      ></div>
-
-                      {/* Content Section */}
-                      <div className="card-content">
-                        <h2 className="card-title">{event.name}</h2>
-                        <div className="space-y-2">
-                          <div className="info-row">
-                            <Calendar className="w-4 h-4 text-cyan-400" />
-                            <span>{event.date}</span>
-                          </div>
-                          <div className="info-row">
-                            <User className="w-4 h-4 text-rose-400" />
-                            <span>{event.organizer}</span>
-                          </div>
-                          <div className="info-row">
-                            <MapPin className="w-4 h-4 text-violet-400" />
-                            <span>{event.location}</span>
-                          </div>
-                          <div className="info-row">
-                            <Tag className="w-4 h-4 text-amber-400" />
-                            <span>{event.event_type}</span>
-                          </div>
-                        </div>
-
-                        {event.url && (
-                          <a
-                            href={event.url}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            className="event-link"
-                          >
-                            View Event <ExternalLink className="w-4 h-4" />
-                          </a>
-                        )}
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </Marquee>
-            </motion.div>
+            <MarqueeSection events={events} />
           </>
         )}
       </div>
@@ -189,65 +129,3 @@ export default function Events() {
     </div>
   );
 }
-
-const EventCard = ({ event }: { event: Event }) => (
-  <div className="event-card">
-    {/* Image Section */}
-    <div>
-      {event.cover_image_url ? (
-        <img
-          src={event.cover_image_url}
-          alt={event.name}
-          className="w-full h-150 object-cover"
-        />
-      ) : (
-        <div className="w-full h-150 bg-gradient-to-r from-slate-800 to-slate-700"></div>
-      )}
-    </div>
-
-    {/* Content Section */}
-    <div className="card-content">
-      <h2 className="card-title">{event.name}</h2>
-      <div className="space-y-2">
-        <div className="info-row">
-          <Calendar className="w-4 h-4 text-cyan-400" />
-          <span>{event.date}</span>
-        </div>
-        <div className="info-row">
-          <User className="w-4 h-4 text-rose-400" />
-          <span>{event.organizer}</span>
-        </div>
-        <div className="info-row">
-          <MapPin className="w-4 h-4 text-violet-400" />
-          <span>{event.location}</span>
-        </div>
-        <div className="info-row">
-          <Tag className="w-4 h-4 text-amber-400" />
-          <span>{event.event_type}</span>
-        </div>
-      </div>
-
-      {event.url && (
-        <a
-          href={event.url}
-          target="_blank"
-          rel="noopener noreferrer"
-          className="event-link"
-        >
-          View Event <ExternalLink className="w-4 h-4" />
-        </a>
-      )}
-    </div>
-  </div>
-);
-const InfoRow = ({ icon, text }: { icon: React.ReactNode; text: string }) => (
-  <div className="info-row">
-    {icon}
-    <span className="truncate">{text}</span>
-  </div>
-);
-
-const truncateText = (text: string, maxLength: number) => {
-  if (text.length <= maxLength) return text;
-  return text.slice(0, maxLength) + "...";
-};
