+++ frontend/src/app/network/NetworkFloatingCard.tsx
