--- frontend/src/components/ui/_modal.css
@@ -1,58 +1,0 @@
-.modal-backdrop {
-  position: fixed;
-  inset: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  backdrop-filter: blur(4px);
-  z-index: 40;
-}
-
-.modal-container {
-  transform-origin: center;
-  will-change: transform;
-}
-
-.modal-content {
-  position: relative;
-  z-index: 1;
-}
-
-.modal-content::before {
-  content: "";
-  position: absolute;
-  inset: 0;
-  background: linear-gradient(
-    to bottom right,
-    rgba(255, 255, 255, 0.1),
-    rgba(255, 255, 255, 0.05)
-  );
-  border-radius: inherit;
-  pointer-events: none;
-}
-
-/* Scrollbar styling */
-.modal-container::-webkit-scrollbar {
-  width: 8px;
-}
-
-.modal-container::-webkit-scrollbar-track {
-  background: rgba(0, 0, 0, 0.1);
-  border-radius: 4px;
-}
-
-.modal-container::-webkit-scrollbar-thumb {
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 4px;
-}
-
-.modal-container::-webkit-scrollbar-thumb:hover {
-  background: rgba(255, 255, 255, 0.3);
-}
-
-/* Card hover effect */
-.modal-slide-up {
-  transition: transform 0.2s ease-out;
-}
-
-.modal-slide-up:hover {
-  transform: translateY(-2px);
-}
